requirements.txt
astroid
autopep8
certifi
charset-normalizer
flake8
idna
isodate
isort
jedi
lazy-object-proxy
lxml
mccabe
mypy
mypy-extensions
pycodestyle
pydocstyle
pyflakes
pylint
pyparsing
python-dateutil
python-jsonrpc-server
python-language-server
python-lsp-jsonrpc
python-lsp-server
pytz
rdflib
rdflib-jsonld
requests
rope
six
snowballstemmer
toml
typing-extensions
ujson
urllib3
wrapt
yapf
web.py
SPARQLWrapper
gitpython
APScheduler==3.10.1
